Command(+ Timer) + Error

+++ TelemetrySoftware2019/src/configuration/ConfReader.java
@@ -221,11 +221,11 @@ public class ConfReader {
 	 * Read and return timer lenght in ms
 	 * On reading error log and return 3000
 	 */
-	public static int getTimerLen(){
+	public static long getTimerLen(){
 		try {
 			JSONObject obj = readJSONObject();
 			JSONObject rec = (JSONObject)obj.get("command");
-			return (int)rec.get("timer");
+			return (long)rec.get("timer");
 		} catch (Exception e) {
 			System.err.println("Config file reading error. Return 3000 as default lenght");
 			return 3000;
